+++ src/main/java/br/com/ottimizza/dashboard/controllers/KpiDetailController.java
@@ -3,6 +3,8 @@ package br.com.ottimizza.dashboard.controllers;
 import br.com.ottimizza.dashboard.models.Kpi;
 import br.com.ottimizza.dashboard.models.KpiDetail;
 import br.com.ottimizza.dashboard.services.KpiDetailService;
+
+import java.math.BigInteger;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -33,7 +35,7 @@ public class KpiDetailController {
         
     @GetMapping("find/{id}")
     //<editor-fold defaultstate="collapsed" desc="Find kpi by id">
-        public ResponseEntity<Optional<KpiDetail>> findKpiByID(@PathVariable("id") Long idKpi) throws Exception{
+        public ResponseEntity<Optional<KpiDetail>> findKpiByID(@PathVariable("id") BigInteger idKpi) throws Exception{
             return ResponseEntity.ok(kpiService.findById(idKpi));
         }
     //</editor-fold>
@@ -48,7 +50,7 @@ public class KpiDetailController {
         
     @DeleteMapping("delete/{id}")
     //<editor-fold defaultstate="collapsed" desc="Delete kpi">
-        public ResponseEntity<String> removeKpi(@PathVariable("id") Long idKpi) throws Exception{
+        public ResponseEntity<String> removeKpi(@PathVariable("id") BigInteger idKpi) throws Exception{
             return ResponseEntity.ok(kpiService.delete(idKpi).toString());
         }
     //</editor-fold>
